requirements.txt
# Server dependencies
py-cord>=2.4.1
yt-dlp>=2024.1.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyNaCl>=1.5.0
